watcher/requirements.txt
google-cloud-storage

python-dotenv
